update mock and move to new location
