+++ include/eeprom/core/book_accessor.hpp
@@ -63,16 +63,17 @@ class BookAccessor
     explicit BookAccessor(
         EEpromTaskClient& eeprom_client, accessor::ReadListener& read_listener,
         DataBufferType<SIZE>& buffer,
-        dev_data::DevDataTailAccessor<EEpromTaskClient>& tail_accessor)
+        dev_data::DevDataTailAccessor<EEpromTaskClient>& tail_accessor,
+        eeprom::CRC16Base& crc16)
         : accessor::EEPromAccessor<EEpromTaskClient,
                                    addresses::ot_library_begin>(
               eeprom_client, *this,
               accessor::AccessorBuffer(intermediate_buffer.begin(),
                                        intermediate_buffer.end())),
           tail_accessor(tail_accessor),
+          crc16(crc16),
           read_listener(read_listener),
           buffer(buffer) {
-        crc16_init();
         eeprom_client.send_eeprom_queue(
             message::ConfigRequestMessage{config_req_callback, this});
     }
@@ -138,6 +139,7 @@ class BookAccessor
     // fields, decide what they are
     // Add a tail accessor?
     dev_data::DevDataTailAccessor<EEpromTaskClient>& tail_accessor;
+    eeprom::CRC16Base& crc16;
     message::ConfigResponseMessage conf = message::ConfigResponseMessage{};
     bool config_updated{false};
     table_entry_action action_cmd_m = dev_data::table_entry_action{};
@@ -148,8 +150,8 @@ class BookAccessor
     template <size_t num_bytes>
     auto calc_crc(std::array<uint8_t, num_bytes> data)
         -> std::array<uint8_t, 2> {
-        uint16_t crc = eeprom::crc16_compute(data.cbegin(),
-                                             static_cast<uint8_t>(num_bytes));
+        uint16_t crc =
+            crc16.crc16_compute(data.cbegin(), static_cast<uint8_t>(num_bytes));
         std::array<uint8_t, 2> crc_byte{};
         std::memcpy(crc_byte.data(), &crc, sizeof(crc));
