+++ include/eeprom/firmware/crc16.h
@@ -1,14 +1,15 @@
 #pragma once
-#include <stdint.h>
+#include <cstdint>
 
 namespace eeprom {
 
-class Crc16 {
+class CRC16Base {
   public:
+    virtual ~CRC16Base() = default;
     /**
      * Initialize crc module.
      */
-    virutal void crc16_init() = 0;
+    virtual void crc16_init() = 0;
 
     /**
      * Compute the CRC
@@ -16,7 +17,7 @@ class Crc16 {
      * @param length Length of data
      * @return Computed CRC
      */
-    uint16_t crc16_compute(const uint8_t* data, uint8_t length);
+    virtual uint16_t crc16_compute(const uint8_t* data, uint8_t length) = 0;
 
     /**
      * Continue accumulating CRC using provided data.
@@ -24,38 +25,41 @@ class Crc16 {
      * @param length Length of data
      * @return Accumulated CRC
      */
-    uint16_t crc16_accumulate(const uint8_t* data, uint8_t length);
+    virtual uint16_t crc16_accumulate(const uint8_t* data, uint8_t length) = 0;
 
     /**
      * Reset the accumulated CRC value.
      */
-    void crc16_reset_accumulator();
+    virtual void crc16_reset_accumulator() = 0;
 };
 
-/**
- * Initialize crc module.
- */
-void crc16_init();
+class CRC16Accelerated : public CRC16Base {
+  public:
+    /**
+     * Initialize crc module.
+     */
+    void crc16_init() override;
 
-/**
- * Compute the CRC
- * @param data Data
- * @param length Length of data
- * @return Computed CRC
- */
-uint16_t crc16_compute(const uint8_t* data, uint8_t length);
+    /**
+     * Compute the CRC
+     * @param data Data
+     * @param length Length of data
+     * @return Computed CRC
+     */
+    uint16_t crc16_compute(const uint8_t* data, uint8_t length) override;
 
-/**
- * Continue accumulating CRC using provided data.
- * @param data Data
- * @param length Length of data
- * @return Accumulated CRC
- */
-uint16_t crc16_accumulate(const uint8_t* data, uint8_t length);
+    /**
+     * Continue accumulating CRC using provided data.
+     * @param data Data
+     * @param length Length of data
+     * @return Accumulated CRC
+     */
+    uint16_t crc16_accumulate(const uint8_t* data, uint8_t length) override;
 
-/**
- * Reset the accumulated CRC value.
- */
-void crc16_reset_accumulator();
+    /**
+     * Reset the accumulated CRC value.
+     */
+    void crc16_reset_accumulator() override;
+};
 
 }  // namespace eeprom
